+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - `MarabouONNXNetwork` no longer has a `shallowCopy` method. Instead of calling this method,
     you should set the new parameter `preserveExistingConstraints` in the method `readONNX` to
     `True` which has the same effect.
+  - The method `getMarabouQuery` on `MarabouNetwork` has been renamed `getInputQuery`.
 
 * Added support for creating constraints using the overloaded syntax `<=`, `==` etc. in
   the Python backend. See `maraboupy/examples/7_PythonicAPI.py` for details.